.github/actions/medium-publish/requirements.txt
requests==2.28.2
python-frontmatter==1.0.0